@@ -11,7 +11,7 @@
     "check": "svelte-check --tsconfig ./tsconfig.json && tsc -p tsconfig.node.json"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@testing-library/jest-dom": "^6.4.8",
     "@testing-library/svelte": "^5.2.1",
     "@testing-library/user-event": "^14.5.2",
@@ -19,12 +19,12 @@
     "autoprefixer": "^10.4.19",
     "jsdom": "^24.1.1",
     "postcss": "^8.4.40",
-    "svelte": "^4.2.18",
-    "svelte-check": "^3.8.4",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tailwindcss": "^3.4.7",
     "tslib": "^2.6.3",
-    "typescript": "^5.2.2",
-    "vite": "^5.3.4",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vitest": "^2.0.4"
   },
   "dependencies": {
