@@ -1,12 +1,17 @@
 <script lang="ts">
   import { cn } from "../utils/cn";
 
-  export let selected: boolean = false;
-  export let text: string;
+  interface Props {
+    selected?: boolean;
+    text: string;
+    onClick: () => void;
+  }
+
+  let { selected = false, text, onClick }: Props = $props();
 </script>
 
 <button
-  on:click
+  onclick={onClick}
   class={cn(
     "py-2 rounded-md hover:bg-light-cyan hover:text-very-dark-cyan",
     selected
